sink, source: Consolidate passthrough setup in reconfigure
This moves over the saving+resetting/restoring of volumes and source
suspending/unsuspending while entering/leaving passthrough mode into
reconfigure functions. This makes it easier to reason about exactly what
behaviour occurs at the time, as well as avoids loss of precision during
the remapping of the internal volume values in this case.

Change-Id: If36a09b394bc64d982124bf7ce035c3c3c8fa306