stream: Add the ability for clients to provide timestamps/durations
This adds infrastructure and API for clients to send per-buffer
timestamps and durations during playback. The idea is that sinks which
can support such timestamp-based rendering can use these to play the
data at the designated time, and potentially use hardware features to
perform synchronisation and rate matching.

We also add durations at this juncture for future work to support
compressed data, as that will need clients to provide per-buffer
durations, and we might as well make the API changes at one go.

Some things to note:

  1) This increases the per-message client-side overhead by almost a
     factor of two, as we are now passing 4 extra 32-bit ints for the
     timestamp. However, this is only true when timestamp/duration
     information is provided, so the impact outside of that use-case is
     minimal.

  2) The hard assumption that is being made is that if _any_ operation
     occurs on the buffer between the client sending the data and the
     sink popping it, the timestamp and duration are invalidated. The
     rationale for this is that clients have requested fine-grained
     control of what buffers are rendered when, and that we should not
     try to modify that. The only exception to this rule is the
     application of vike, That said,  We may be able to relax this for
     PCM streams, but not for compressed streams (as we cannot do the
     same sort of bytes->time calculation in that case).

As an implementation note, the protocol changes as follows:

  * In SHM mode, we add the timing information after the SHM
    information in the packet payload

  * In non-SHM mode, we add the timing information after the pstream
    descriptors and before the frame data payload

Change-Id: I9de3d2b10d4ee36d3428892b19e5ed17bff31f5f