qahw: Modify logic to avoid concurrent loopback and capture
Modify logic to ensure that source is moved to suspended state
when loopback starts. Once loopback session is stopped,
source is resumed.
For dynamic sources, source is resumed after checking for any
config changes. Incase of config change, source is removed and
added with new config. For static sources, source is directly
resumed without and check.

Change-Id: Ia8dfd5eedff9a96d5e29d8c8e4bed0c95c7f7a37