tests: Add a unit test for timestamping support
We add a "timestamp_mode" to module-null-sink which merely takes
timestamp'ed buffers from a client (starting from 0), and validates that
the incoming timestamps and durations correctly correspond with the data
that is coming in.

We then use this in a unit test to make sure that timestamps/durations
are correctly passed fro clients to the sink without an invalidation en
route.

Change-Id: Id2fe3ae6dced5a3f3c08522d1b567d5577722c5c