sink-input,source-output: Always reconfigure while entering passthrough
As we now do some amount of bookkeeping for entering/leaving the
passthrough state, let's always call reconfigure to keep track of when
entering the passthrough state. This is especially required because we
run reconfigure unconditionally when leaving passthrough, and at that
point if we don't find saved state, we assert out.

Change-Id: Id519583e796623604f8093d3c087746eb6985b2e