null-sink: Add a compressed mode for testing
This adds support for rendering compressed data to null-sink. The
purpose is to make it possible to test compressed output without having
actual hardware that can consume this data.

We still need to add a way to validate that the data we get is actually
untouched and not broken up by anything in the core.

Change-Id: I1e5260f6a54e039d36ab3a64d9bda0e2e222cd50