combine-sink: Perform latency computations even if adjust_time is 0
This is needed to make sure we get accurate latency reports even when
the rate adjustment mechanism is turned off. It might make sense to
provide a different know for the latency calculations, or just to move
that out into its own separate timer event, but this is less intrusive
for now.

Change-Id: I939d80d29089411363fbe2884a660a5b3b4e3f98