sink: Add a drain method for compressed streams
We cannot simple complete a drain as we do with PCM streams for
compressed streams as there might be a fair amount of data buffered in
the compressed sink. To account for this, we add an *asynchronous* drain
method on the sink. This should trigger a drain, and on completion, the
sink should call pa_sink_drain_complete() which in turn will notify the
stream that the drain was actually completed.

Change-Id: I9e8fc9fed0cd54b70b12a120de12a6e2cbb0b114