sink: Add convenience API to render an unbroken memblock
For sinks that need to work with compressed or timestamped data, we
expect that only one sink-input would be connected at any given point of
time, and that the memblock being popped should not be broken up (as
that might invalidate timestamp/duration data, or provide an incomplete
compressed frame). This adds API to achieve this without having sinks
check for silence memblocks, or worry about the requested length being
big enough. The expectation is that the sink will handle any partial
writes if the buffer is too large.

Change-Id: I3641283be626e624878a9b25fcc0a7accfd21433