stream: Add the ability to provide per-buffer flags
This extends the pa_stream_write_ts*() API to also allow for providing
buffer flags. Currently, we only support a flag to signal a
discontinuous buffer, but presumably we can find other uses for this.

From a protocol perspective, there is no extra cost -- we stuff the
buffer flags into some unused bits in the pstream descriptor flags.

We still need to find a good way to test this.

Change-Id: Ic4417da33fa5e16cf26f68b1bb034d4bee2af78d