bluetooth: Update logic for Device association to Adapter
In current logic all the devices are tried to associate to adapter
only once, and cant be attempted because of the flad used even when
the adapter is available.

Updated the logic to associate and validate the devices upon addition
of a Adapter and always validate a new device and associate to
an existing adapter.

Change-Id: If2729ade4a3ea20ab88ea916008cc41e8bbd92b8