Use pa_nsec_t as the type for PA_*NSEC* constants
Those constants are supposed to be used with pa_nsec_t values. As-is, on
some 64-bit platform, pa_nsec_t is not the same type as the constants
(unsigned long vs unsigned long long) causing unnecessary warnings from
compilers/IDEs.

Also, this is more consistent with how the matching USEC constants are
defined.

Change-Id: I2da703a9bd9e3f0ceaf31dbd8765f4964799a890