Swap AUX and TOP channels as needed
PulseAudio only provides channels for what is typically called "height
speakers", which are above the front and rear speakers, using the "TOP"
positions.
Dolby's 7.1.4 expects "top speakers", which are half-way between the
listener and the front/rear speakers. PulseAudio doesn't have named
channels for those so we use the AUX[0-3] channels instead.

Our soundbar/multichannel support only support "height/TOP" speakers so
Dolby's "top" audio is lost.
As a quick workaround, swap between "TOP" and "AUX" channel based on what
is provided and what is wanted, similar to what is done for side vs rear
channels in a 5.1 setup.

Change-Id: I36664975f47cd471b737984eb28bb93b4da30458