Add a PA_SINK_NO_REMIX_OVERRIDE flag to sinks
This flag can be used to notify the upstream sink-input that remixing
is not wanted. Remixing will now be disable if either the sink-input or
the sink don't want it.

This is useful when it's not possible to disable remixing in the
sink-input, e.g. slaves of combine-sink, or if the sink itself has
particular needs, e.g. doing channel filtering in remap-sink.

Also update remap-sink to support that flag via a "remix_override"
parameter, and remove the NO_REMIX flag from ts_renderer since that was
a workaround to disable remixing for a downstream remap-sink.

Change-Id: I9ea862976c487c79106df0278cd127a8c66a6f60