Remove unnecessary call to getLatency
The getLatency call is quite expensive (200-500us) but unless we push
new data, there is no reason for the resulting playback_time to change
(any decrease in latency should have a matching increase in the current
time).

(Also set a project name for the top-level CMakeLists file for IDEs that
can load it directly, e.g. Qt Creator)

Change-Id: I16f6ce0080be48d9ec23bc74dea926c512d1ac70