group-modules: don't require timestamp and duration
Don't require a valid timestamp for all the packets in a timestamp stream so we
can handle packet getting split (it's not necessary to have a timestamp for
every packet to get a good synchronization. And if upstream needs to split a
packet, it is better to not pass any timestamp than passing an incorrect one)

Similarly, duration is not required. It's mainly used to compute the current
end of a stream (last packet TS + last packet duration), in which case the
number of samples gives an accurate enough duration.

Change-Id: I3e2ec586fb1f4562cfb9bec65f1f26471ca70ef3