qal: Update channel map correctly for source.
Port configuration info is used to fill channel map info for
qal device and source configuration info is used to fill channel
map info for qal stream. In the existing code, port configuration
was used to update channel map for stream & source configuration
was used to update channel map for device which is incorect.
Hence fixed it.

Change-Id: I1d38ed9d4305baa9bc5c3186c92bc02992f40496