Reset timestamp computation on underruns
For non-timestamp stream, when there is an underrun, reset the timestamp
computation, otherwise we can get into a loop of underruns (underrun ->
wait for buffer to fill -> read from the buffer quickly to catch up from
the lost time -> underrun).

However, we can't just reset the timestamp immediately because it's used
to compute the sink latency, which is some times used by upstream to
estimate where we are in the stream (stream clock), which is some times
used to know when to trigger a EOS in the media app. If we reset the
timestamp, the latency will become the target latency and upstream will
believe that there is that much yet to play before the EOS.

Change-Id: Ibbd0ed93607e1ba37454beafd5f587717d530b99