Core: On pop_one failure do not set playing_for to 0
playing_for is sent to client in every timing info update. And timing
info at the client is only updated when the server reports that it is
playing. On the same lines underrun_for, reports how long it has
underrun.

In peek_one, since we don't return slience, towards the end of stream we
immediately set playing_for to zero, but we keep on asking the client for
more data with playing_for = 0. Client interpolation logic see that the
server is no longer playing and it stops updating the timing info.
This results in an infinite wait on the client.