group_manager: don't reset timestamp computation when ahead of target
At the start of stream, the latency might be higher than targeted because
downstream is trying to fill its buffer.

For instance, if the target latency is 40ms because downstream's buffer
is 40ms, at the beginning of the stream, the first packet will be:
	<now>+40ms
and the last packet in the buffer will be:
	<now>+40ms(latency of first packet)+40ms(size of the buffer)
So the last timestamp will be now+80ms, well ahead of the target latency.

In that case, if we have an underrun, there is no need to reset the
timestamp computation.

Change-Id: I99c324907a014be30d0b09f9437e286e4199837c