group-manager: combine too-small chunks
When a chunk is smaller than a threshold (currently 5ms), and doesn't
have a timestamp, combine it with next chunk(s), for downstream sinks
that have a minimum chunk size.

For chunks with timestamp, rely on the app to provide the size.

This issue doesn't apply in regular PulseAudio since a sink can use
pa_sink_render_full to enforce the size requirement.

Change-Id: Ic7e1cffa7f98f53a46ea66fd7eb06fe63d8093f1