Native Sampling Rate support
Implemented sink reconfiguration to change the sample rate to the desired rate
and format, and avoid resampling.
However the resampler may still be invoked in case of upmix/downmix
of the channels based on the sink graph.
Below are the changes for remap-sink, combine-sink, group-manager, group-sink :
1. Changed the condition check in pa_sink_reconfigure when
avoid_process is true - set desired spec rate of the sink
same as the rate of the sink-input/stream if they are not equal
2. Defined new callbacks for sink->reconfigure in remap-sink,
combined-sink, group-manager
3. In remap-sink and group-manager, moved the code for
static sink input creation into independant functions to
be able to reuse at the time of module init as well as sink reconfigure.
4. Added support to accept param 'avoid_processing' as a module argument
 for remap-sink, combined-sink, group-manager and group-sink(depends on the changes
done in pa_config_update and 100-base pre-init scripts)
5. Modified module-null-sink.c to update the avoid_processing value of
'data' after its init.

Change-Id: Iab3af083bf8578aea49eaa910a1c9e3b91f3a789