Non-split A2dp stream timestamping in pulseaudio
Use a sliding window to figure out when the packets could arrive at the
earliest (i.e. if they were not delayed by the BT transport). Use that
to figure out the latency. Then apply a correction to the packet duration
to maintain fixed latency (i.e. slow down or speed up the playback to
maintain the latency)

Also add a Dbus method to dynamically set ttp offset and properties to
controle the sliding window.

CRs-Fixed: 2815163
Change-Id: I0097ce6db098a9ebab77c363fa559b7795a95291