Initialize ts_log struct for A2DP source instance.
While ts_log is used only for A2DP sinks, it's part of the struct even
when in A2DP source mode.

If not initialized, the ts_log file descriptor was set to 0, which is
used for some other descriptor (typically a PA message queue). When a BT
stream was tear down, it was then closing the wrong file descriptor, which
later caused an assertion error in the PA message queue.

Change-Id: I568801be4861610604a26a95c40fe9cd32066242