pulseaudio: add the ability to get the value of pulseaudio version from yocto
If this git repo does not have a tag or has a tag exception, then
the vaule of pulseaudio version obtained from the "git describe"
command by calling git-version-gen in meson.build would be null,
which would result in a build error.

So, avoid this issue by receiving the correct version value from yocto.

Change-Id: Iac38900f2e9f0a165a5d30a90e0c9733af374335