module-qal-card: fix double free issue on 'jack_in_config'
'jack_in_config' is allocated for port hdmi-out, and pass it
to port usb-in. so while exit pulseaudio, port hdmi-out first
to free itself momery includes 'jack_in_config', and then port
usb-in same to free memory, it dected 'jack_in_config' is not
null and then free it, caused double free in 'jack_in_config'.

Change-Id: If599957065399616566077b7e34f537e30fbbdef